fix binmg bong
